Implement GDB stub for remote debugging (#45)
With the help of the external mini-gdbstub library, rv32emu offers an
experimental way to support a limited number of GDB Remote Serial
Protocol (GDBRSP) packets. After applying this patch, we were able to
connect rv32emu to GDB and interact with it using basic commands
such as "print $pc" to dump the value of current program counter or
"continue" to resume program running.

It is important to note that this patch only permits the use of one
breakpoint. Future upgrades will be made.

Close #22