Resolve bug in red-black tree (#72)
Even though the commit 6161d53d62cbb13446747bbbbbd134d6d651b5e3 has
shortened the map code, I found that it may be incorrect because the
second parameter passed to the function "rb_set_parent" on line 537 is
"*indirect", which was just assigned to the value of new node in the
previous statement. I added a variable "parent" to store the previous
value of "indirect" before any changes were made and set it as the
parent of "new_node" as the old code intended.