Decouple RV32 instructions as implementation template
This commit lays the foundation for the forthcoming development of
a foundational JIT compiler. Through the separation of RV32[IMACF]
instructions, reusable implementation templates are established,
providing a basis for code generation within the JIT compiler.

This separation allows for the creation of flexible and direct code
generation strategies, without any redundancy between the interpreter
and the baseline JIT compiler. Notably, macro operation fusion
instructions are omitted from the template, as they are intended to
appear within the optimizer module.