CI: Implementing benchmark regression tests (#213)
The 'pull_request_target' event is utilized for the benchmark action due
to its ability to update data in the 'gh-pages' branch, which is crucial
for visualization on GitHub Pages. The 'pull_request' event lacks the
necessary 'GITHUB_TOKEN' for this task.

Furthermore, 'workflow_dispatch' event is added to enable the initial
setup and running of the benchmark. This permits the storage of base
benchmark data for future comparisons.

To prevent redundant executions, a filter has been implemented to
exclude the merging push event.

Close #166