Add appropriate 'const' qualifier to the parameters
The 'const' qualifier is useful for ensuring type safety. Specifically,
'const' can clarify pointer ownership. It is important to note that
'const' applied to the target of a pointer or reference does not imply
that the target is constant. Instead, it indicates a read-only view of
the object. This means that while the object can not be modified through
this particular reference or pointer, it may still be altered through
other means. An object is truly immutable only when 'const' is used in
its definition. The primary purpose of using 'const' is not to facilitate
compiler optimizations, but to safeguard against programming errors.