Improve the performance of indirect jump for T1C
Considering the T1C ends with the indirect jump, our interpreter mode
now records the target of the indirect jump, identifying the most
frequent jump target for T1C. The T1 generated code for the indirect
jump compares the selected jump target and executes the jump if the
comparison results in equality. With this enhancement, T1C and T2C
which are integrated later can proceed with an indirect jump if the
target PC matches the selected jump target.

Based on the performance analysis, the benchmarks with significant
number of indirect jump effectively improve the performance.

| Metric   | N_in_jmp  | Original | Propused | Speedup |
|----------+-----------+----------+----------+---------|
|miniz	   |    2098313|   1.266 s|   1.225 s|   +3.35%|
|sha512    |   10500727|   1.905 s|   1.861 s|   +2.36%|
|dhrystone |   20000618|   0.325 s|   0.253 s|  +28.46%|
|nqueens   |   44658722|   1.051 s|    0.79 s|  +33.04%|
|qsort	   |  275000250|   1.978 s|   1.517 s|  +30.39%|