Bind I/O handlers during emulator initialization
It is not required to give an application the opportunity to bind I/O
handlers because I/O handlers are rarely altered during the
creation of an emulator.

With this commit, the application can now build a emulator much more
easier by only taking the emulator's attribute (vm_attr_t)
into consideration.

In order to facilitate further integration with the RISC-V system
emulator (semu), I have included a TODO inside the I/O interface.

Related: #310