CI: Add Clang static analyzer (#361)
The clang static analyzer is invoked by executing scan-build. 

If any issues are detected by scan-build, it will halt the pipeline. 
If halting the pipeline is not the desired behavior, remove the flag 
--status-bugs .

The analysis result can be viewed in the GitHub Action execution log.

Ubuntu-latest is used to enable us to utilize the latest version of the 
tool.

References:
- https://gitlab.xfce.org/xfce/xfce4-dev-tools/-/issues/51
- https://gitlab.com/gnuwget/wget2/-/blob/cabedf3847afd3926cbeaa67d4ee9f28f78979d9/.gitlab-ci.yml#L279