Modularize the extractor and verifier for external
Abstract the extractor and verifier to support .zip and .gz compressed
files.

The $(call) function executes during the expansion, so we have to make
all steps inside download-extract-verify to $(call) like, so that they
can be run in sequentially. To do that, introducing three more functions
which are prologue, download, and epilogue.