Fix wasm build failed in deploy-wasm CI
Before executing the "make CC=emcc ENABLE_GDBSTUB=0 ENABLE_SDL=1"
command during the deploy-wasm ci, the $(BIN) must be modified to
$(BIN).js. We can easily fix this by moving the "include" mk statement
ahead of the "all" target. After reordering, the first invoked make
target is not "all" target but instead when no make target is specified,
set the .DEFAULT_GOAL to "all" to solve it.