Consolidate consistency of AMO RVOP (#428)
The 'addr' variable is used to specify the memory address in many I/O
operations within the RVOP macro, such as the 'lh' and 'lw' instructions.
However, most of the AMO (Atomic Memory Operation) instructions did not
use it. To maintain consistency, the use of the 'addr' variable has been
consolidated.