Add supervisor memory fence instruction SFENCE.VMA (#458)
The supervisor memory-management fence instruction SFENCE.VMA is used
to synchronize updates to in-memory memory-management data structures
with current execution. Linux kernel leverage this instruction for
synchronization, support emulating this instruction is needed to boot
Linux kernel. Since we have only one hart and we have not yet support
cache (TLB) at present, the implementation of this instruction is
straight forward and it could be further improved in the future.

Related: #310, #438