Add Wait for Interrupt instruction (WFI)
The Wait for Interrupt instruction (WFI) provides a hint to the
implementation that the current hart can be stalled until an
interrupt might need servicing. As well, Linux kernel might call
wait_for_interrupt() to enter idle state. In this commit, we
could simply make PC + 4 and not really enter idle state.

Related: #310