Ask to install RISCOF before running arch-test locally
If someone attempts to run make arch-test instead of executing the
script ".ci/riscv-tests.sh", the target will bail because riscof
is not the default Python module. This alert can easily clarify
any confusion.

A specific version of RISCOF is used in CI, so requesting for that
version as well.