ERROR
Attempt to refactor the operation in `V*_LOOP`. Since the original
implementation is too stiff to handle all the instructions in RVV.

Notice the it fails for now where I assume all the `V*_LOOP` macros are
correct but when defining the operation function the compiler fail to
link the operation function.